Fix type assignment in singleton case
Because we re-use Cell instance, we have to set type correctly.

In original design, we only set self.type if type is not string.

However, if we re-use the instance, the self.type may be :integer but
we may not assign self.type value even `cell_type_from_value` return
:string type.

Original design is to save memory, and the current implementation does
not violate the objective, because in the newly created Cell instance,
which does not have self.type defined, self.type will be :string and
the assignment will not performed since `self.type != type`.